Hent aktive deltakelser for tiltak (iff admin i tiltaksadmin)
